% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irced.r
\name{add_nick}
\alias{add_nick}
\title{Add nickname/user info}
\usage{
add_nick(irc_obj, nickname = "keryx", username = "keryx",
  realname = "keryx", password = NULL)
}
\arguments{
\item{nickname}{nickname (default is "\code{keryx}")}

\item{username}{username (default is "\code{keryx}")}

\item{realname}{realname (default is "\code{keryx}")}

\item{password}{auth password for user (NOTE: not implemented yet)}
}
\description{
You only need to use this function if you want to change the
nick/user/real names or need to authenticate the nick. The default
nickmame (et al) is \href{https://www.britannica.com/topic/keryx}{keryx}.
}
\examples{
irc_server("irc.rud.is") \%>\%
  set_channel("#builds") \%>\%
  add_nick("nick", "nick", "nick") \%>\%
  post_message("Finishing a build is important, but building is more important.")
}

