% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irced.r
\name{bot_connect}
\alias{bot_connect}
\title{Make bot connection}
\usage{
bot_connect(irc_obj, handler)
}
\arguments{
\item{irc_obj}{irc object}

\item{handler}{bot handler R function}
}
\description{
Make bot connection
}

