% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{irc_notify}
\alias{irc_notify}
\title{Send a message/notification to an IRC server/channel}
\usage{
irc_notify(irc_server, channel = "#", message = "", port = 6667L,
  ssl = FALSE, server_password = "", nickname = "RMsgBot",
  username = "RMsgBot", realname = "RMsgBot")
}
\arguments{
\item{irc_server}{IRC server FQDN or IP address}

\item{channel}{channel on the IRC server}

\item{message}{character vector of messages to send}

\item{port}{IRC server port}

\item{ssl}{use secure connection to IRC server?}

\item{server_password}{server password or \code{NULL}}

\item{nickname}{nickname (default is "\code{RMsgBot}")}

\item{username}{username (default is "\code{RMsgBot}")}

\item{realname}{realname (default is "\code{RMsgBot}")}
}
\description{
Given an IRC server and port, plus channel and character vector,
send the contents of the character vector as separate messages to
the specified channel.
}
\note{
Due to inherent limitations in the IRC protocol, each message must not
      exceed 493 characters. Furthermore, since IRC messages are CRLF terminated,
      messages should not include these characters.
}
\examples{
irc_notify("irc.rud.is", "#builds", "test")
}

