% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irced-package.R
\docType{package}
\name{irced}
\alias{irced}
\alias{irced-package}
\title{Putting the R into IRC}
\description{
\code{irced} : \\ˈɜrkt\\\cr
\cr
verb\cr
\cr
1. what users in your IRC channels will become
if you send too many automated messages.
}
\details{
Build systems such as 'Travic-CI' have many ways to notify
users when jobs have completed including posting a message to an IRC
channel. In similar form, there are many times when analyses
or data cleaning/processing operations are long running tasks that
operate in the background or on remote systems. Functions are provided
to notify users of the status of these tasks via IRC.
}
\author{
Bob Rudis (@hrbrmstr)
}

