% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irced.r
\name{post_message}
\alias{post_message}
\title{Post an IRC message}
\usage{
post_message(irc_obj, channel, message, channel_password = NULL)
}
\arguments{
\item{message}{character vector of messages to send}
}
\description{
Post an IRC message
}
\note{
Due to inherent limitations in the IRC protocol, each message must not
      exceed 493 characters. Furthermore, since IRC messages are CRLF terminated,
      messages should not include these characters.
}
\examples{
irc_server("irc.rud.is") \%>\%
  post_message("#builds", "Finishing a build is important, but building is more important.")
}

