% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irced.r
\name{set_channel}
\alias{set_channel}
\title{Set channel info}
\usage{
set_channel(irc_obj, channel, password = NULL)
}
\arguments{
\item{channel}{channel to join on the IRC server}

\item{password}{channel password}
}
\description{
Set channel info
}
\examples{
irc_connect("irc.rud.is") \%>\%
  set_channel("#builds") \%>\%
  post_message("Finishing a build is important, but building is more important.")
}

