% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irced.r
\name{add_nick}
\alias{add_nick}
\title{Add nickname/user info}
\usage{
add_nick(irc_obj, nickname = "keryx", username = "keryx",
  realname = "keryx", password = NULL)
}
\arguments{
\item{irc_obj}{irc object}

\item{nickname}{nickname (default is "\code{keryx}")}

\item{username}{username (default is "\code{keryx}")}

\item{realname}{realname (default is "\code{keryx}")}

\item{password}{auth password for user \emph{(NOTE: not implemented yet)}.
If your server requires \href{https://freenode.net/kb/answer/sasl}{SASL}
authentication then specify this information in the call to
\code{irc_connect()} as noted in the help for that function. Use this
method for the \code{/msg NickServ IDENTIFY foo password} method of
user authentication to IRC. Be wary of using any authentication
scheme without an SSL connection to the IRC server.}
}
\description{
You only need to use this function if you want to change the
nick/user/real names or need to authenticate the nick. The default
nickmame (et al) is \href{https://www.britannica.com/topic/keryx}{keryx}.
}
\examples{
irc_server("irc.rud.is") \%>\%
  add_nick("nick", "nick", "nick") \%>\%
  irc_connect() -> session
}

