% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irced.r
\name{irc_connect}
\alias{irc_connect}
\title{Create an IRC server object}
\usage{
irc_connect(host_or_ip, port = 6667, password = NULL, use_ssl = FALSE,
  verbose = FALSE)
}
\arguments{
\item{host_or_ip}{IRC server FQDN or IP address}

\item{port}{IRC server port}

\item{password}{server password or \code{NULL}. If you are required to
do \href{https://freenode.net/kb/answer/sasl}{SASL} authentication then
this should be "\code{nick}:\code{password}". Be wary of using any
authentication scheme without an SSL connection to the IRC server.}

\item{use_ssl}{use secure connection to IRC server?}

\item{verbose}{produce \emph{very} chatty IRC diagnostic messages?}
}
\description{
Create an IRC server object
}
\examples{
irc_connect("irc.rud.is") \%>\%
  set_channel("#builds") \%>\%
  post_message("Finishing a build is important, but building is more important.")
}

