% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irced.r
\name{irc_server}
\alias{irc_server}
\title{Create an IRC server object}
\usage{
irc_server(host_or_ip, port = 6667, password = NULL, use_ssl = FALSE)
}
\arguments{
\item{host_or_ip}{IRC server FQDN or IP address}

\item{port}{IRC server port}

\item{password}{server password or \code{NULL}}

\item{use_ssl}{use secure connection to IRC server?}
}
\description{
Create an IRC server object
}
\examples{
irc_server("irc.rud.is") \%>\%
  set_channel("#builds") \%>\%
  post_message("Finishing a build is important, but building is more important.")
}

